requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
redis>=5.0.0
torch>=2.1.0
transformers>=4.36.0
pyyaml>=6.0
prometheus-client>=0.19.0
numpy>=1.24.0
